Share the directory walk between the commands that need it
`file add` grew an in-process file walk built on the `ignore` crate, which is
the walker fd itself uses. A second caller is about to want the same thing, so
move it to `walk.rs` rather than have two copies drift apart, and move the
file-contents preview to `pick.rs`, where previews live.

Also honour `.fdignore` while there: users who keep one for fd expect the same
paths to stay out of scriv's pickers.
